// frontend/src/types/ServerInfoResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ServerInfoResponse = { id: number, ip: string, online: number, max: number, version_name: string, protocol: number, license: boolean, disconnect_reason_html: string | null, updated: string, description_html: string, was_online: boolean, is_checked: boolean, is_spoofable: boolean | null, is_crashed: boolean, is_forge: boolean, favicon: string | null, ping: bigint | null, };